created mosaic components
